Allow metadata preservation

+++ argus/recipes/cloud/base.py
@@ -201,4 +201,8 @@ class BaseCloudbaseinitRecipe(base.BaseRecipe):
         self.get_cb_init_logs()
         self.get_cb_init_confs()
 
-        self.delete_mock_metadata()
+        if CONFIG.argus.delete_metadata:
+            LOG.info("Deleting metadata.")
+            self.delete_mock_metadata()
+        else:
+            LOG.info("The metadata was preserved.")
